Add C-style cast Egor sorry

+++ src/Game/GameState.cpp
@@ -94,7 +94,7 @@ void GameState::handle(const AttackEvent &event) {
     Unit *attacker = findUnitByID(event.attackerID_);
     Unit *target = findUnitByID(event.targetID_);
     if(!attacker || !target) {
-        std::cerr << "Ruslan za chto " << event.attackerID_ << ' ' << event.targetID_ << std::endl;
+        std::cerr << "Ruslan za chto " << (int)event.attackerID_ << ' ' << (int)event.targetID_ << std::endl;
     }
     for (auto & attack : attacker->myAttacks_) {
         if (!attack->getMovingFlag()) {
